+++ changelog.md
@@ -1,3 +1,6 @@
+latest
+ - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.
+
 2024.05.14 1.4.14
  - Added basic error catch in the joinAll method to help surface chain build errors.
  - User alert for userAgent version checks to inform users if their browser is of a minimum supported version for the browser API's used.